개별적으로 사용되는 enum inner class로 전환. example 패키지 삭제"
