+++ src/main/java/was/common/entity/BaseTimeEntity.java
@@ -14,13 +14,10 @@ import java.time.LocalDateTime;
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
 public abstract class BaseTimeEntity {
-
     @CreatedDate
     @Column(updatable = false)
     private LocalDateTime created;
 
     @LastModifiedDate
     private LocalDateTime updated;
-
-
 }
