--- src/main/java/was/example/domain/member/Member.java
@@ -1,4 +1,0 @@
-package was.example.domain.member;
-
-public class Member {
-}
